Update patch set 3
Patch Set 3: Code-Review-1

This will not help as kolla is removing enabled config files from both locations:

https://github.com/openstack/kolla/blob/master/docker/horizon/Dockerfile.j2#L42-L43

In fact current rpm is already shipping enabled config files in both:

# rpm -qlp openstack-manila-ui-2.10.2-0.20170904151218.57c34ea.el7.centos.noarch.rpm |grep '_9090'
/usr/lib/python2.7/site-packages/manila_ui/local/enabled/_9090_manila_admin_add_share_group_types_panel_to_share_panel_group.py
/usr/lib/python2.7/site-packages/manila_ui/local/enabled/_9090_manila_admin_add_share_group_types_panel_to_share_panel_group.pyc
/usr/lib/python2.7/site-packages/manila_ui/local/enabled/_9090_manila_admin_add_share_group_types_panel_to_share_panel_group.pyo
/usr/share/openstack-dashboard/openstack_dashboard/local/enabled/_9090_manila_admin_add_share_group_types_panel_to_share_panel_group.py
/usr/share/openstack-dashboard/openstack_dashboard/local/enabled/_9090_manila_admin_add_share_group_types_panel_to_share_panel_group.pyc
/usr/share/openstack-dashboard/openstack_dashboard/local/enabled/_9090_manila_admin_add_share_group_types_panel_to_share_panel_group.pyo

IMO the problem is in kolla. I think some (or all) of the plugins are automatically disabled when there is not endpoint for the services, although i'm not sure if this applies to all of them.

Patch-set: 3
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1