Add configs for manila-ui under enabled
In horizon we have two different ways to configure
our dashboards: with enabled and with local_settings.d.
With enable we can toggle panels on and off, while configs
on local_settings.d are used to override configs.

We were copying configs under local_settings.d but we were
missing configs under enabled.

This patch set makes sure to get that config files in place.

Change-Id: I18351aff7fdc860534795e844bec419ea76c54ef