Update patch set 1
Patch Set 1:

Thanks Victoria for the patch :)

To give a bit of context to reviewers, we've had an issue in kolla where the enable scripts were not in the expected location, causing manila-ui panel to not show up.

By copying the files instead of moving them, this makes openstack-manila-ui packaging more consistent with the other horizon plugins and fixes the kolla issue.

https://bugs.launchpad.net/kolla/+bug/1774175

Patch-set: 1