Add symlink control for manila-ui configuration files
There is a difference between RDO python-django-horizon
and RHOS python-django-horizon, precisely in upstream
the local config folder is not a symlink so we need to
create symlinks for each of the configs, whereas in
downstream the local config folder is a symlink and
we need to get rid of the symlinking for the files,
just installing them in place is enough.

Setting up a flag to control this will make the import
process easier and less error-prone.

Change-Id: I4f5977695eb508a52e5f9922aa40420a8913a6bd